+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * feat: support for immutable adder and remover
 * feat: optional second argument for getting the existing target value
 * refactor: separate dynamic properties determination to dedicated class
+* refactor: add an abstraction for property read/write info extractor
 
 ## 1.8.0